Rename source code file for Ch. 9
For some reason, git is not recognizing the rename and modified file in
a single commit as it has in previous commits, causing the file history
to be lost. As a workaround, I will rename the file in this commit, then
make the changes for Chapter 9 in the following commit.